start adding tests

+++ programs/chapter-x-nft/src/constants/period.rs
@@ -1,1 +1,1 @@
-pub const STAKE_PERIOD_IN_DAYS: u64= 20;
+pub const STAKE_PERIOD_IN_DAYS: f64 = 20.0;
